requirements.txt
Flask
Gunicorn
Pillow
numpy
werkzeug